requirements.txt
# Install PyTorch GPU wheels from the official CUDA 12.8 index.
# JAX is intentionally kept on the CPU build in this project.
--extra-index-url https://download.pytorch.org/whl/cu128

# essential
torch==2.7.0+cu128
tabpfn==7.1.1
jax==0.9.2
jaxlib==0.9.2
jaxtyping==0.3.9
numpy==2.4.4
equinox==0.13.6
chex==0.1.91
tqdm==4.67.3
typing-extensions>=4.0

# for configuration management
hydra-core==1.3.2

# for downloading and splitting datasets
openml==0.15.1
ucimlrepo==0.0.7
scikit-learn==1.8.0
pandas==3.0.2
matplotlib==3.10.8
